Rename functional tests to deterministic tests
